add an argument to the click handler function
when the browser calls an event handler, whether it is a click
or any kind of handler, it also passes an argument.

this argument contains the even object, which contains target
and currentTarget properties.

Use the debugger in the browser developer tools to peek into
the argument and check both properties.