log the currentTarget property of the event object
do the same with the currentTarget property of the event object
as we did with the target property.

now when the user clicks on the first list item in the page
while the console is open, we are able to see the target and
the currentTarget of the event which are passed to the event
handler function by the browser.