add few inline semantic tags to the list item
inline semantic tags such as <em>, <strong> and <abbr> can
be used to emphasize and elaborate on specific words and
phrases within a paragraph or a list item.

the fact these semantic tags apply to portions of the
paragraph (one or more words) without effecting the entire
paragraph means that they are inline tags or correspond to
inline nodes in the document flow.

these tags will allow us to test our click handler further
and dig deeper into the difference between the target and the
currentTarget properties of the event argument to the handler
function.