replace lorem bullets with empty list placeholder
the empty placeholder and the <ul> tags are given id's to make
it easier to retirieve them from javascript using the document.
querySelector method.

the click handler function is kept which will be used for
actual todo items

for the empty actual item, which is not an actual todo item,
the DOM node is retrieved and stored in a constant called
nodeEmpty, which will have to have its visibility toggled
depending on whether we have items or not in the list.

this is the next baby step...