create a function update empty item
the idea of this function is to show or hide the empty item
placeholder based on whether or not the list is empty

it is needed because there are two situations in which the
placeholder needs to be shown or hidden:
- when the first actual todo item is added to the list
- when the last actual todo item is removed from the list

so instead of replicating the logic both in the item click
handler and in the button click handler, it is a better idea
to create this function and call it from both places

in the current implementation, the function just **attempts**
to hide the empty placeholder

_but this does NOT work_

the bug is intricate and is hard to discover, this is your
next challenge, is to try to debug the code and examine it in
the browser to come up with a theory as to why this is not
happening

Try to call the toggleEmptyItem manually with true and false
before and after adding a todo item. what do you notice?