fix difficult bug from previous baby step
so what happened and why were things not working as expected?

the answer is not so obvious. since nodeTodoList was defined
once and assigned the value of the DOM node found by searching
for an <li> node with an id of "empty", each time we rewrote
the innerHTML property (in the addTodo function) the constant
previously referenced in nodeEmpty was destroyed and recreated
making the global constant valid, but no longer a node that
still exists in the DOM

tricky, isn't it?

how would someone troubleshoot or diagnose a problem like this?
first you need to checkout the previous commit, run it in the
browser and check the value of nodeEmpty in the console

you will notice that before adding a todo, this element when
typed in the console highlights the empty placeholder item in
the list

however and after something is added using addTodo and the list
is rewritten, and even though the constant still holds the
value of what seems to be a DOM element, typing the name no
longer highlights the empty item in the list

a careful examiner would then try to use querySelector again
to obtain the empty placeholder item in the list and compare
it with nodeEmpty using strict equality (===)

this will make it clear that what we currently have is another
node, and that our beloved empty placeholder item had been
murdered and replaced with another clone 😥

so how do we fix this? the easiest fix is to requery the node
each time to obtain the most up to date DOM node inside of
the toggleEmptyItem function

this is possible since this is the only place in the code
currently where this constant is needed, otherwise we might
have found ourselves in a place where this simple fix is not
the best thing to do but it is for the time being!

for the next baby step we would like to find a way to show
the empty placeholder item when the last todo item is removed