simplify updateEmptyElement
since we are passing true to updateEmptyElement when isEmpty
is true and passing false to updateEmptyElement when isEmpty
is false, a refactoring idea is to eliminate the if statement
and simply pass isEmpty to updateEmptyElement as an argument.

clever, isn't it?

soon you will be able to see these possibilities naturally,
we are just baby programmers 🍼 here so let's continue with
the baby steps!

the next baby step is to maybe change the message when the last
todo item is removed from "nothing to do" to "you finished!" or
similar, to motivate the user